fix migration

+++ database/migrations/2024_01_23_121849_create_article_trade_names_table.php
@@ -12,8 +12,8 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('article_trade_names', function (Blueprint $table) {
-            $table->integer('article_id')->unsigned();
-            $table->integer('trade_name_id')->unsigned();
+            $table->bigInteger('article_id')->unsigned();
+            $table->bigInteger('trade_name_id')->unsigned();
             $table->foreign('article_id')->references('id')->on('articles')->onUpdate('cascade')->onDelete('cascade');
             $table->foreign('trade_name_id')->references('id')->on('ultrashop_trade_names')->onUpdate('cascade')->onDelete('cascade');
             $table->unique(['article_id', 'trade_name_id']);
